Avoid restarts of swift-proxy, memcached (bnc#776903)
The mechanism is
http://wiki.opscode.com/display/chef/Resources#Resources-Notifications

Instead of unconditionally restarting swift-proxy and memcached,
only restart them if their configs change.

Use "service subscribes to config" instead of "config notifies
service" because the immediate restart needs to happen only when the
config has been written.

action :start is needed if the service dies for any reason. Normally
it is started by the above notification when configured initially.