Respect order of crowbar orchestration in recipes
The orchestration of the swift deployment is tricky, and crowbar does
several steps:
 - swift-storage (to discover disks)
 - ring-compute (to compute the rings)
 - swift-storage (to setup swift daemons)
 - swift-proxy
 - swift-dispersion

All of this starts falling apart (and causing failures) if a node has
several roles. So in each recipe, we also check the step we're in to
decide whether to go on with the recipe.

The way we identify the steps:
 - swift-storage (to discover disks): we have the role, but have not
   found the devices yet in the compile phase (works because
   swift-storage compile phase is run after the other recipes)
 - ring-compute: we have the role, and the rings are not on disk yet in
   the compile phase