Do not setup swift-proxy if rings are not present
If the rings are not present, then swift-proxy will just fail to start.

If we do not have the rings yet in the execute phase when trying to setup the
services, then it means the ring-compute node didn't push them yet or we didn't
manage to pull them yet. Since we know that swift-proxy will fail without the
rings, there's no point in doing anything.

This can happen for various reasons, but the most likely one is having the
periodic chef-client run occur on the swift-proxy node before the crowbar
chef-client run occurs on the swift-ring-compute node. We know that crowbar
will run chef-client on the swift-proxy node after the rings have been
computed, so we will have the rings and a chance to start swift-proxy later on.

https://bugzilla.novell.com/show_bug.cgi?id=872518